fixing problems with authorization, updating changelog and readme
