+++ graphql/schema.graphql
@@ -1,0 +1,41 @@
+type Query {
+  getTaxConfiguration: TaxConfiguration
+}
+
+type Mutation {
+  setTaxConfiguration(operation: String): OrderFormConfiguration
+}
+
+type OrderFormConfiguration {
+  paymentConfiguration: PaymentConfiguration
+  taxConfiguration: TaxConfiguration
+  minimumQuantityAccumulatedForItems: Int
+  decimalDigitsPrecision: Int
+  minimumValueAccumulated: Int
+  apps: [App]
+  allowMultipleDeliveries: Boolean
+  allowManualPrice: Boolean
+  maxIntOfWhiteLabelSellers: Int
+  maskFirstPurchaseData: Boolean
+  recaptchaValidation: Boolean
+}
+
+type PaymentConfiguration {
+  requiresAuthenticationForPreAuthorizedPaymentOption: Boolean
+  allowInstallmentsMerge: Boolean
+  blockPaymentSession: Boolean
+  paymentSystemToCheckFirstInstallment: Boolean
+}
+
+type TaxConfiguration {
+  allowExecutionAfterErrors: Boolean
+  authorizationHeader: String
+  integratedAuthentication: Boolean
+  url: String
+}
+
+type App {
+  fields: [String]
+  id: [String]
+  major: Int
+}
