+++ manifest.json
@@ -8,6 +8,7 @@
   "categories": [],
   "dependencies": {},
   "builders": {
+    "graphql": "1.x",
     "node": "6.x",
     "docs": "0.x"
   },
@@ -31,6 +32,13 @@
       "attrs": {
         "host": "api.vtex.com",
         "path": "/api/logistics/*"
+      }
+    },
+    {
+      "name": "outbound-access",
+      "attrs": {
+        "host": "api.vtex.com",
+        "path": "/api/catalog/*"
       }
     }
   ],
