+++ node/package.json
@@ -6,7 +6,7 @@
     "@types/co-body": "^0.0.3",
     "@types/jest": "^24.0.18",
     "@types/node": "^12.0.0",
-    "@vtex/api": "6.35.1",
+    "@vtex/api": "6.36.0",
     "@vtex/test-tools": "^1.0.0",
     "tslint": "^5.14.0",
     "tslint-config-vtex": "^2.1.0",
